+++ docs/reference/google.auth.transport.aiohttp_requests.rst
@@ -1,7 +1,7 @@
 google.auth.transport.aiohttp\_requests module
 ==============================================
 
-.. automodule:: google.auth.transport.aiohttp_requests
+.. automodule:: google.auth.transport._aiohttp_requests
    :members:
    :inherited-members:
    :show-inheritance:
