+++ docs/reference/google.auth.transport.rst
@@ -12,7 +12,7 @@ Submodules
 .. toctree::
    :maxdepth: 4
 
-   google.auth.transport.aiohttp_requests
+   google.auth.transport._aiohttp_requests
    google.auth.transport.grpc
    google.auth.transport.mtls
    google.auth.transport.requests
