+++ tests_async/test_jwt_async.py
@@ -20,7 +20,7 @@ import pytest
 
 from google.auth import crypt
 from google.auth import exceptions
-from google.auth import jwt_async
+from google.auth import _jwt_async as jwt_async
 from tests import test_jwt
